move WorkFLowRevisionCount from package mongo/entity to mongo/model
