+++ src/main/java/io/boomerang/service/crud/WorkflowVersionServiceImpl.java
@@ -24,11 +24,11 @@ import io.boomerang.model.projectstormv5.TaskNode;
 import io.boomerang.mongo.entity.FlowTaskTemplateEntity;
 import io.boomerang.mongo.entity.FlowUserEntity;
 import io.boomerang.mongo.entity.RevisionEntity;
-import io.boomerang.mongo.entity.WorkFlowRevisionCount;
 import io.boomerang.mongo.entity.WorkflowEntity;
 import io.boomerang.mongo.model.ChangeLog;
 import io.boomerang.mongo.model.Dag;
 import io.boomerang.mongo.model.Revision;
+import io.boomerang.mongo.model.WorkFlowRevisionCount;
 import io.boomerang.mongo.model.WorkflowScope;
 import io.boomerang.mongo.model.next.DAGTask;
 import io.boomerang.mongo.service.FlowTaskTemplateService;
